Created new DAL-parser to parse the data based on client used.
