+++ src/index.js
@@ -2,8 +2,7 @@ const crypto = require('crypto');
 const {firstOrNull, isEmpty, arrOrDefault} = require('./utils');
 const KnexClient = require('../knex/knex-client');
 const {ballotsTable, multisigMembershipsTable, blocksTable, accountsTable, delegatesTable, transactionsTable, storeTable} = require('../knex/ldpos-table-schema');
-const parsers = require('./parser');
-
+const DalParser = require("./parsers");
 const DEFAULT_NETWORK_SYMBOL = 'ldpos';
 const ID_BYTE_SIZE = 20;
 
@@ -12,6 +11,7 @@ class DAL {
     config = config || {};
     this.logger = config.logger || console;
     this.knexClient = new KnexClient(config);
+    this.parsers = new DalParser(this.knexClient).getAppliedParsers()
   }
 
   async init(options) {
@@ -532,7 +532,7 @@ class DAL {
   }
 
   repository(tableName, ...primaryKeys) {
-    const dataReadParser = parsers[tableName];
+    const dataReadParser = this.parsers[tableName];
     const basicRepositoryOps = (defaultMatcher) =>
       ({
         get: (equalityMatcher = defaultMatcher) => this.knexClient.findMatchingRecords(tableName, equalityMatcher, dataReadParser),
